+++ devtools/gradle/gradle-extension-plugin/src/main/java/io/quarkus/extension/gradle/tasks/ExtensionDescriptorTask.java
@@ -7,7 +7,6 @@ import java.io.InputStream;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
@@ -39,6 +38,7 @@ import io.quarkus.bootstrap.BootstrapConstants;
 import io.quarkus.bootstrap.model.AppArtifactCoords;
 import io.quarkus.bootstrap.model.AppArtifactKey;
 import io.quarkus.bootstrap.model.ApplicationModelBuilder;
+import io.quarkus.devtools.project.extensions.ScmInfoProvider;
 import io.quarkus.extension.gradle.QuarkusExtensionConfiguration;
 import io.quarkus.extension.gradle.dsl.Capability;
 import io.quarkus.extension.gradle.dsl.RemovedResource;
@@ -311,7 +311,7 @@ public class ExtensionDescriptorTask extends DefaultTask {
     }
 
     private void computeSourceLocation(ObjectNode extObject) {
-        Map<String, String> repo = getSourceRepo();
+        Map<String, String> repo = ScmInfoProvider.getSourceRepo();
         if (repo != null) {
             ObjectNode scm = extObject.putObject("scm");
             for (Map.Entry<String, String> e : repo.entrySet()) {
@@ -319,22 +319,6 @@ public class ExtensionDescriptorTask extends DefaultTask {
 
             }
         }
-    }
-
-    static Map<String, String> getSourceRepo() {
-        // We could try and parse the .git/config file, but that will be fragile
-        // Let's assume we only care about the repo for official-ish builds produced via github actions
-        String repo = System.getenv("GITHUB_REPOSITORY");
-        if (repo != null) {
-            Map info = new HashMap();
-            String qualifiedRepo = "https://github.com/" + repo;
-            // Don't try and guess where slashes will be, just deal with any double slashes by brute force
-            qualifiedRepo = qualifiedRepo.replace("github.com//", "github.com/");
-
-            info.put("url", qualifiedRepo);
-            return info;
-        }
-        return null;
     }
 
     private void computeQuarkusCoreVersion(ObjectNode extObject) {
