+++ independent-projects/extension-maven-plugin/src/main/java/io/quarkus/maven/ExtensionDescriptorMojo.java
@@ -68,6 +68,7 @@ import io.quarkus.bootstrap.resolver.maven.MavenArtifactResolver;
 import io.quarkus.bootstrap.resolver.maven.workspace.LocalProject;
 import io.quarkus.bootstrap.resolver.maven.workspace.LocalWorkspace;
 import io.quarkus.bootstrap.util.DependencyUtils;
+import io.quarkus.devtools.project.extensions.ScmInfoProvider;
 import io.quarkus.fs.util.ZipUtils;
 import io.quarkus.maven.capabilities.CapabilitiesConfig;
 import io.quarkus.maven.capabilities.CapabilityConfig;
@@ -574,22 +575,6 @@ public class ExtensionDescriptorMojo extends AbstractMojo {
         return originalArtifact.replace("${project.version}", projectVersion);
     }
 
-    static Map<String, String> getSourceRepo() {
-        // We could try and parse the .git/config file, but that will be fragile
-        // Let's assume we only care about the repo for official-ish builds produced via github actions
-        String repo = System.getenv("GITHUB_REPOSITORY");
-        if (repo != null) {
-            Map info = new HashMap();
-            String qualifiedRepo = "https://github.com/" + repo;
-            // Don't try and guess where slashes will be, just deal with any double slashes by brute force
-            qualifiedRepo = qualifiedRepo.replace("github.com//", "github.com/");
-
-            info.put("url", qualifiedRepo);
-            return info;
-        }
-        return null;
-    }
-
     private static JsonNode getJsonElement(ObjectNode extObject, String... elements) {
         JsonNode mvalue = extObject.get(elements[0]);
         int i = 1;
@@ -689,7 +674,7 @@ public class ExtensionDescriptorMojo extends AbstractMojo {
     }
 
     private void addSource(ObjectNode extObject) throws MojoExecutionException {
-        Map<String, String> repo = getSourceRepo();
+        Map<String, String> repo = ScmInfoProvider.getSourceRepo();
         if (repo != null) {
             ObjectNode scm = extObject.putObject("scm");
             for (Map.Entry<String, String> e : repo.entrySet()) {
